refactor: move connector config in the relative connector file
