+++ tests/test_codemanager.py
@@ -6,8 +6,11 @@ import uuid
 import pandas as pd
 import pytest
 
-from pandasai.connectors.base import SQLConnectorConfig
-from pandasai.connectors.sql import PostgreSQLConnector, SQLConnector
+from pandasai.connectors.sql import (
+    PostgreSQLConnector,
+    SQLConnector,
+    SQLConnectorConfig,
+)
 from pandasai.exceptions import BadImportError, NoCodeFoundError
 from pandasai.helpers.skills_manager import SkillsManager
 from pandasai.llm.fake import FakeLLM
